feat : auth middleware

+++ internal/model/user.go
@@ -3,7 +3,7 @@ package model
 import "time"
 
 type User struct {
-	Id        int64     `json:"id" validate:"-"`
+	Id        int       `json:"id" validate:"-"`
 	Username  string    `json:"username" validate:"email" example:"example@mail.org"`
 	Password  string    `json:"password,omitempty" validate:"omitempty,min=6,max=30" example:"pass12345678"`
 	CreatedAt time.Time `json:"-" validate:"-"`
